Use tile-level labels in a PyTorch dataloader
- `Dataset.torch()` supports tile-level labels in the form of a pandas dataframe. The dataframe is expected to be indexed by `{slide}-{loc_x}-{loc_y}` and have a column named "label".
- Support training with tile-level labels, by directly instancing a `sf.model.Trainer` and providing a pandas DataFrame to the `labels` argument, in the same format listed above. Only support single outcomes (multiple categories, or single continuous outcome).
- Various bug fixes
